產品create

+++ app/Http/Controllers/ProductTypeController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\productType;
+use App\ProductType;
 use Illuminate\Http\Request;
 
 class ProductTypeController extends Controller
@@ -16,26 +16,26 @@ class ProductTypeController extends Controller
 
     }
     public function index(){
-        $product_types = productType::get();
+        $product_types = ProductType::get();
         return view($this->index,compact('product_types'));
     }
     public function create(){
-        $types = productType::TYPE;
+        $types = ProductType::TYPE;
         return view($this->create,compact('types'));
 
     }
     public function store(Request $request){
-        productType::create($request->all());
+        ProductType::create($request->all());
         return redirect('/admin/product/type')->with('message', '新增成功');
 
     }
     public function edit($id){
-        $record = productType::find($id);
-        $types = productType::TYPE;
+        $record = ProductType::find($id);
+        $types = ProductType::TYPE;
         return view($this->edit,compact('record','types'));
     }
     public function update(Request $request,$id){
-        $old_record = productType::find($id);
+        $old_record = ProductType::find($id);
         $old_record->type = $request->type;
         $old_record->type_name = $request->type_name;
         $old_record->save();
@@ -44,7 +44,10 @@ class ProductTypeController extends Controller
     }
     public function delete($id)
     {
-        $old_record = productType::find($id);
+        $old_record = ProductType::find($id);
+        if($old_record->products->count()!=0){
+            return redirect('/admin/product/type')->with('message', '仍有 '.$old_record->product->count().' 筆商品，無法刪除該產品種類。');
+        }
         $old_record->delete();
 
         return redirect('/admin/product/type')->with('message', '刪除成功');
